capstone_project/app/requirements.txt
opencv-contrib-python-headless==4.2.0.34
matplotlib==3.1.3
Keras==2.4.3
streamlit==0.66.0
numpy==1.18.1
tensorflow-cpu==2.3.0
pandas==1.0.1
Pillow==7.2.0
